+++ homeworks/max-hirning/21-next-js-1/src/styles/Home.module.css
@@ -1,7 +1,24 @@
+.header {
+
+}
+
+.nav {
+  height: 5rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+}
+
+.link {
+  font-size: 2rem;
+  font-weight: bold;
+  font-family: sans-serif;
+}
+
 .main {
   display: flex;
   padding: 1rem;
-  min-height: 100vh;
+  min-height: 89vh;
   align-items: center;
   flex-direction: column;
   justify-content: space-between;
